integration changes to report on master
fix include scenario section in variables

fix include scenario section in operate

add some logs

add some logs

add some logs

add some logs

add some logs

add some logs

+++ app/repositories/BranchRepository.scala
@@ -68,7 +68,9 @@ class BranchRepository @Inject()(db: Database) {
 
   def deleteAll(branches: Seq[Branch]): Unit = {
     db.withConnection { implicit connection =>
-      SQL"DELETE FROM branch WHERE id IN (${branches.map(_.id)})".executeUpdate()
+      if (branches.nonEmpty){
+         SQL"DELETE FROM branch WHERE id IN (${branches.map(_.id)})".executeUpdate()
+      }
       ()
     }
   }
